requirements.txt
ccxt>=4.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
PyYAML>=6.0
